+++ src/lib/iTin.Core.Hardware/iTin.Core.Hardware.Specification/iTin.Core.Hardware.Specification.Smbios/SMBIOS/Structures/Specific/GroupAssociationElement.cs
@@ -35,7 +35,7 @@ namespace iTin.Core.Hardware.Specification.Smbios
 
         #region private properties
 
-        #region [private] (int) Handle: Gets a value that represents the 'Handle' field
+        #region [private] (ushort) Handle: Gets a value that represents the 'Handle' field
         /// <summary>
         ///  Gets a value that represents the '<b>Handle</b>' field.
         /// </summary>
@@ -43,7 +43,7 @@ namespace iTin.Core.Hardware.Specification.Smbios
         /// Value of the property.
         /// </value>
         [DebuggerBrowsable(DebuggerBrowsableState.Never)]
-        private int Handle => GetWord(0x01);
+        private ushort Handle => (ushort)GetWord(0x01);
         #endregion
 
         #region [private] (SmbiosStructure) ItemType: Gets a value that represents the 'ItemType' field
